+renamed telegram automation to telegramPosting +added data and photos, +added keyboards
