+++ Python/telegramPosting/autoPosting/handlers/__init__.py
@@ -1,1 +1,2 @@
 import handlers.start
+import handlers.menu
