+++ Python/telegramPosting/autoPosting/requirements.txt
