+++ Python/telegramPosting/autoPosting/utils/methods.py
@@ -2,8 +2,8 @@ import asyncio
 import aioschedule
 
 
-async def add_task(func):
-    aioschedule.every(5).seconds.do(func)
+async def add_task(func, text='Content', photo=None):
+    aioschedule.every(5).seconds.do(func, text, photo)
 
     while True:
         await aioschedule.run_pending()
